Refactored whole calculations: transferred logic to domain object, used BigDecimals

+++ src/main/java/pl/jt/demo/hotelincomecalculator/infra/requests/CalculationRequest.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public record CalculationRequest(int premiumRoomsAvailable,
                                  int economyRoomsAvailable,
-                                 List<Integer> willingnessToPayList) {
+                                 List<Double> willingnessToPayList) {
 }
